+++ .gitignore
@@ -11,3 +11,4 @@ node_modules
 .eslintcache
 .vscode/**
 .DS_Store
+.env.local
