--- packages/nextjs/person.json
@@ -1,1 +1,0 @@
-{"name":"John","age":30,"email":"john.doe@example.com"}
